Sample job script + Call Script
These two scripts are examples of how to run the Blender algorithm. For memory reasons, the runs were conducted in batches of ~200 pixels. So, to run all of them I used ~20 copies of samplejob.txt that each call RunTuolumne.sh starting at a different pixel value. The in-line comment in samplejob.txt indicates where the start number is supplied to RunTuolumne.sh